Refactor conversation builder helpers
